Moved cell value formatting logic into formatters
